person_project/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
